part1: new serialization impl in py
