+++ python/pyfury/__init__.py
@@ -33,7 +33,6 @@ if ENABLE_FURY_CYTHON_SERIALIZATION:
         Language,
         ClassInfo,
         OpaqueObject,
-        ComplexObjectSerializer,
     )
 else:
     from pyfury._fury import (  # noqa: F401,F403,F811 # pylint: disable=unused-import
@@ -42,9 +41,9 @@ else:
         ClassInfo,
         OpaqueObject,
     )
-    from pyfury._struct import (  # noqa: F401,F403,F811 # pylint: disable=unused-import
-        ComplexObjectSerializer,
-    )
+from pyfury._struct import (  # noqa: F401,F403,F811 # pylint: disable=unused-import
+    ComplexObjectSerializer,
+)
 from pyfury.serializer import *  # noqa: F401,F403 # pylint: disable=unused-import
 from pyfury.type import (  # noqa: F401 # pylint: disable=unused-import
     record_class_factory,
