+++ python/pyfury/tests/test_cross_language.py
@@ -444,7 +444,7 @@ class ComplexObject2:
 
 def test_serialize_simple_struct_local():
     fury = pyfury.Fury(language=pyfury.Language.XLANG, ref_tracking=True)
-    fury.register_class(ComplexObject2, type_tag="test.ComplexObject2")
+    fury.register_type(ComplexObject2, type_tag="test.ComplexObject2")
     obj = ComplexObject2(f1=True, f2={-1: 2})
     new_buf = fury.serialize(obj)
     assert fury.deserialize(new_buf) == obj
@@ -453,7 +453,7 @@ def test_serialize_simple_struct_local():
 @cross_language_test
 def test_serialize_simple_struct(data_file_path):
     fury = pyfury.Fury(language=pyfury.Language.XLANG, ref_tracking=True)
-    fury.register_class(ComplexObject2, type_tag="test.ComplexObject2")
+    fury.register_type(ComplexObject2, type_tag="test.ComplexObject2")
     obj = ComplexObject2(f1=True, f2={-1: 2})
     struct_round_back(data_file_path, fury, obj)
 
@@ -461,8 +461,8 @@ def test_serialize_simple_struct(data_file_path):
 @cross_language_test
 def test_serialize_complex_struct(data_file_path):
     fury = pyfury.Fury(language=pyfury.Language.XLANG, ref_tracking=True)
-    fury.register_class(ComplexObject1, type_tag="test.ComplexObject1")
-    fury.register_class(ComplexObject2, type_tag="test.ComplexObject2")
+    fury.register_type(ComplexObject1, type_tag="test.ComplexObject1")
+    fury.register_type(ComplexObject2, type_tag="test.ComplexObject2")
 
     obj2 = ComplexObject2(f1=True, f2={-1: 2})
     obj1 = ComplexObject1(
@@ -504,7 +504,7 @@ def test_serialize_opaque_object(data_file_path):
         data_bytes = f.read()
     debug_print(f"len {len(data_bytes)}")
     fury = pyfury.Fury(language=pyfury.Language.XLANG, ref_tracking=True)
-    fury.register_class(ComplexObject1, type_tag="test.ComplexObject1")
+    fury.register_type(ComplexObject1, type_tag="test.ComplexObject1")
     new_obj = fury.deserialize(data_bytes)
     debug_print(new_obj)
     assert new_obj.f2 == "abc"
@@ -519,12 +519,6 @@ def test_serialize_opaque_object(data_file_path):
 
 
 class ComplexObject1Serializer(pyfury.serializer.Serializer):
-    def get_xtype_id(self):
-        return pyfury.type.FuryType.FURY_TYPE_TAG.value
-
-    def get_xtype_tag(self):
-        return "test.ComplexObject1"
-
     def write(self, buffer, value):
         self.xwrite(buffer, value)
